Splited solution into Main program/interface and class
