+++ TGL_Practice_1/IWork.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TGL_Practice_1
+{
+    interface IWork
+    {       
+        void cutTimber();
+        void rest();
+    }
+}
